+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## (IN PROGRESS)
 
+* Vendor ref number search is not working in the POL. Refs UIOR-668.
+* Update labels for inventory interactions. Refs UIOR-665.
+* Use "Total expended" and "Total encumbered" from composite order. Refs UIOR-658.
+* Update view for Vendor reference number in POL list. Refs UIOR-662.
 * Fix order templates list not refreshed after create new template. Refs UIOR-660.
 * Fix cannot save template with renewal date. Refs UIOR-659.
 * Make POL vendor reference number and type repeatable, paired fields. Refs UIOR-519.